Move public

+++ public/script.js
@@ -2,7 +2,9 @@ $(document).ready(function() {
   $('.user-reviews').on('click', '.trash-can', function() {
     const reviewId = $(this).parent().attr('id')
 
-    deleteReview(reviewId)
+    if (confirm('Are you sure you want to delete this post?')) {
+      deleteReview(reviewId)
+    }
   })
 })
